build: upgrade PHPStan to level 9 and fix strict type errors

+++ src/Rules/ColorContrast.php
@@ -99,14 +99,14 @@ class ColorContrast extends AbstractRule
     {
         $hex = ltrim($hex, '#');
         if (strlen($hex) === 3) {
-            $r = hexdec(substr($hex, 0, 1) . substr($hex, 0, 1));
-            $g = hexdec(substr($hex, 1, 1) . substr($hex, 1, 1));
-            $b = hexdec(substr($hex, 2, 1) . substr($hex, 2, 1));
+            $r = (int)hexdec(substr($hex, 0, 1) . substr($hex, 0, 1));
+            $g = (int)hexdec(substr($hex, 1, 1) . substr($hex, 1, 1));
+            $b = (int)hexdec(substr($hex, 2, 1) . substr($hex, 2, 1));
             return [$r, $g, $b];
         } elseif (strlen($hex) === 6) {
-            $r = hexdec(substr($hex, 0, 2));
-            $g = hexdec(substr($hex, 2, 2));
-            $b = hexdec(substr($hex, 4, 2));
+            $r = (int)hexdec(substr($hex, 0, 2));
+            $g = (int)hexdec(substr($hex, 2, 2));
+            $b = (int)hexdec(substr($hex, 4, 2));
             return [$r, $g, $b];
         }
         return null;
